+++ server/src/jmh/java/org/allaymc/server/BlockStateUpdateJMHTest.java
@@ -30,12 +30,12 @@ public class BlockStateUpdateJMHTest {
 
     @Benchmark
     public void test1() {
-        observer = observer.setPropertyValue(BlockPropertyTypes.MINECRAFT_FACING_DIRECTION, MinecraftFacingDirection.UP);
+        observer = observer.withPropertyValue(BlockPropertyTypes.MINECRAFT_FACING_DIRECTION, MinecraftFacingDirection.UP);
     }
 
     @Benchmark
     public void test2() {
-        observer = observer.setPropertyValue(BlockPropertyTypes.MINECRAFT_FACING_DIRECTION, MinecraftFacingDirection.UP);
-        observer = observer.setPropertyValue(BlockPropertyTypes.POWERED_BIT, true);
+        observer = observer.withPropertyValue(BlockPropertyTypes.MINECRAFT_FACING_DIRECTION, MinecraftFacingDirection.UP);
+        observer = observer.withPropertyValue(BlockPropertyTypes.POWERED_BIT, true);
     }
 }