+++ server/src/main/java/org/allaymc/server/block/component/BlockEndRodBaseComponentImpl.java
@@ -34,7 +34,7 @@ public class BlockEndRodBaseComponentImpl extends BlockBaseComponentImpl {
             face = face.opposite();
         }
 
-        blockState = blockState.setPropertyValue(FACING_DIRECTION, face.ordinal());
+        blockState = blockState.withPropertyValue(FACING_DIRECTION, face.ordinal());
         dimension.setBlockState(placeBlockPos.x(), placeBlockPos.y(), placeBlockPos.z(), blockState, placementInfo);
         return true;
     }